moving plotConvertedReactor to utils
