+++ armi/reactor/converters/geometryConverters.py
@@ -30,11 +30,8 @@ import collections
 import copy
 import math
 import operator
-import os
 from typing import TYPE_CHECKING, Union
 
-import matplotlib
-import matplotlib.pyplot as plt
 import numpy as np
 
 from armi import materials, runLog
@@ -939,7 +936,7 @@ class HexToRZThetaConverter(GeometryConverter):
 
     def plotConvertedReactor(self, fNameBase=None):
         """
-        Generate plots for the converted RZT reactor.
+        Generate plots for the converted RZT reactor. A pass-through to preserve the API.
 
         Parameters
         ----------
@@ -947,125 +944,7 @@ class HexToRZThetaConverter(GeometryConverter):
             A name that will form the basis of the N plots that are generated by this method. Will get split on
             extension and have numbers added. Should be like ``coreMap.png``.
         """
-        runLog.info(f"Generating plot(s) of the converted {str(self.convReactor.core.geomType).upper()} reactor")
-        figs = []
-        colConv = matplotlib.colors.ColorConverter()
-        colGen = plotting.colorGenerator(5)
-        blockColors = {}
-        thetaMesh, radialMesh, axialMesh = self._getReactorMeshCoordinates()
-        innerTheta = 0.0
-        for i, outerTheta in enumerate(thetaMesh):
-            fig, ax = plt.subplots(figsize=(12, 12))
-            innerRadius = 0.0
-            for outerRadius in radialMesh:
-                innerAxial = 0.0
-                for outerAxial in axialMesh:
-                    b = self._getBlockAtMeshPoint(
-                        innerTheta,
-                        outerTheta,
-                        innerRadius,
-                        outerRadius,
-                        innerAxial,
-                        outerAxial,
-                    )
-                    blockType = b.getType()
-                    blockColor = self._getBlockColor(colConv, colGen, blockColors, blockType)
-                    if blockColor is not None:
-                        blockColors[blockType] = blockColor
-                    blockPatch = matplotlib.patches.Rectangle(
-                        (innerRadius, innerAxial),
-                        (outerRadius - innerRadius),
-                        (outerAxial - innerAxial),
-                        facecolor=blockColors[blockType],
-                        linewidth=0,
-                        alpha=0.7,
-                    )
-                    ax.add_patch(blockPatch)
-                    innerAxial = outerAxial
-                innerRadius = outerRadius
-            ax.set_title(
-                "{} Core Map from {} to {:.4f} revolutions".format(
-                    str(self.convReactor.core.geomType).upper(),
-                    innerTheta * units.RAD_TO_REV,
-                    outerTheta * units.RAD_TO_REV,
-                ),
-                y=1.03,
-            )
-            ax.set_xticks([0.0] + radialMesh)
-            ax.set_yticks([0.0] + axialMesh)
-            ax.tick_params(axis="both", which="major", labelsize=11, length=0, width=0)
-            ax.grid()
-            labels = ax.get_xticklabels()
-            for label in labels:
-                label.set_rotation(270)
-            handles = []
-            labels = []
-            for blockType, blockColor in blockColors.items():
-                line = matplotlib.lines.Line2D([], [], color=blockColor, markersize=15, label=blockType)
-                handles.append(line)
-                labels.append(line.get_label())
-            ax.set_xlabel("RADIAL MESH (CM)", labelpad=20)
-            ax.set_ylabel("AXIAL MESH (CM)", labelpad=20)
-            if fNameBase:
-                root, ext = os.path.splitext(fNameBase)
-                fName = root + f"{i}" + ext
-                plt.savefig(fName)
-                plt.close()
-            else:
-                figs.append(fig)
-            innerTheta = outerTheta
-
-        return figs
-
-    def _getReactorMeshCoordinates(self):
-        thetaMesh, radialMesh, axialMesh = self.convReactor.core.findAllMeshPoints(applySubMesh=False)
-        thetaMesh.remove(0.0)
-        radialMesh.remove(0.0)
-        axialMesh.remove(0.0)
-        return thetaMesh, radialMesh, axialMesh
-
-    def _getBlockAtMeshPoint(self, innerTheta, outerTheta, innerRadius, outerRadius, innerAxial, outerAxial):
-        for b in self.convReactor.core.iterBlocks():
-            blockMidTh, blockMidR, blockMidZ = b.spatialLocator.getGlobalCoordinates(nativeCoords=True)
-            if (blockMidTh >= innerTheta) and (blockMidTh <= outerTheta):
-                if (blockMidR >= innerRadius) and (blockMidR <= outerRadius):
-                    if (blockMidZ >= innerAxial) and (blockMidZ <= outerAxial):
-                        return b
-        raise ValueError(
-            "No block found between ({}, {}), ({}, {}), ({}, {})\nLast block had TRZ= {} {} {}".format(
-                innerTheta,
-                outerTheta,
-                innerRadius,
-                outerRadius,
-                innerAxial,
-                outerAxial,
-                blockMidTh,
-                blockMidR,
-                blockMidZ,
-            )
-        )
-
-    @staticmethod
-    def _getBlockColor(colConverter, colGenerator, blockColors, blockType):
-        nextColor = None
-        if blockType not in blockColors:
-            if "fuel" in blockType:
-                nextColor = "tomato"
-            elif "structure" in blockType:
-                nextColor = "lightgrey"
-            elif "radial shield" in blockType:
-                nextColor = "lightgrey"
-            elif "duct" in blockType:
-                nextColor = "grey"
-            else:
-                while True:
-                    try:
-                        nextColor = next(colGenerator)
-                        colConverter.to_rgba(nextColor)
-                        break
-                    except ValueError:
-                        continue
-        return nextColor
+        return plotting.plotConvertedRZTReactor(self.convReactor, fNameBase)
 
     def reset(self):
         """Clear out attribute data, including holding the state of the converted reactor core model."""
