+++ armi/utils/plotting.py
@@ -48,7 +48,7 @@ from armi.reactor import grids
 from armi.reactor.components import Circle, DerivedShape, Helix
 from armi.reactor.components.basicShapes import Hexagon, Rectangle, Square
 from armi.reactor.flags import Flags
-from armi.utils import hexagon, iterables
+from armi.utils import hexagon, iterables, units
 
 LUMINANCE_WEIGHTS = np.array([0.3, 0.59, 0.11, 0.0])
 
@@ -1524,3 +1524,143 @@ def plotConvertedBlock(sourceBlock, convertedBlock, fName=None):
         plt.show()
 
     return fName
+
+
+def plotConvertedRZTReactor(reactor, fNameBase=None):
+    """
+    Generate plots for the converted RZT reactor.
+
+    Parameters
+    ----------
+    fNameBase : str, optional
+        A name that will form the basis of the N plots that are generated by this method. Will get split on extension
+        and have numbers added. Should be like ``coreMap.png``.
+    """
+    runLog.info(f"Generating plot(s) of the converted {str(reactor.core.geomType).upper()} reactor")
+    figs = []
+    colConv = matplotlib.colors.ColorConverter()
+    colGen = colorGenerator(5)
+    blockColors = {}
+    thetaMesh, radialMesh, axialMesh = _getReactorMeshCoordinates(reactor)
+    innerTheta = 0.0
+    for i, outerTheta in enumerate(thetaMesh):
+        fig, ax = plt.subplots(figsize=(12, 12))
+        innerRadius = 0.0
+        for outerRadius in radialMesh:
+            innerAxial = 0.0
+            for outerAxial in axialMesh:
+                b = _getBlockAtMeshPoint(
+                    reactor,
+                    innerTheta,
+                    outerTheta,
+                    innerRadius,
+                    outerRadius,
+                    innerAxial,
+                    outerAxial,
+                )
+                blockType = b.getType()
+                blockColor = _getBlockColor(colConv, colGen, blockColors, blockType)
+                if blockColor is not None:
+                    blockColors[blockType] = blockColor
+                blockPatch = matplotlib.patches.Rectangle(
+                    (innerRadius, innerAxial),
+                    (outerRadius - innerRadius),
+                    (outerAxial - innerAxial),
+                    facecolor=blockColors[blockType],
+                    linewidth=0,
+                    alpha=0.7,
+                )
+                ax.add_patch(blockPatch)
+                innerAxial = outerAxial
+            innerRadius = outerRadius
+        ax.set_title(
+            "{} Core Map from {} to {:.4f} revolutions".format(
+                str(reactor.core.geomType).upper(),
+                innerTheta * units.RAD_TO_REV,
+                outerTheta * units.RAD_TO_REV,
+            ),
+            y=1.03,
+        )
+        ax.set_xticks([0.0] + radialMesh)
+        ax.set_yticks([0.0] + axialMesh)
+        ax.tick_params(axis="both", which="major", labelsize=11, length=0, width=0)
+        ax.grid()
+        labels = ax.get_xticklabels()
+        for label in labels:
+            label.set_rotation(270)
+        handles = []
+        labels = []
+        for blockType, blockColor in blockColors.items():
+            line = matplotlib.lines.Line2D([], [], color=blockColor, markersize=15, label=blockType)
+            handles.append(line)
+            labels.append(line.get_label())
+
+        ax.set_xlabel("RADIAL MESH (CM)", labelpad=20)
+        ax.set_ylabel("AXIAL MESH (CM)", labelpad=20)
+        if fNameBase:
+            root, ext = os.path.splitext(fNameBase)
+            fName = root + f"{i}" + ext
+            plt.savefig(fName)
+            plt.close()
+        else:
+            figs.append(fig)
+        innerTheta = outerTheta
+
+    return figs
+
+
+def _getReactorMeshCoordinates(reactor):
+    """A helper for plotConvertedRZTReactor."""
+    thetaMesh, radialMesh, axialMesh = reactor.core.findAllMeshPoints(applySubMesh=False)
+    thetaMesh.remove(0.0)
+    radialMesh.remove(0.0)
+    axialMesh.remove(0.0)
+    return thetaMesh, radialMesh, axialMesh
+
+
+def _getBlockAtMeshPoint(reactor, innerTheta, outerTheta, innerRadius, outerRadius, innerAxial, outerAxial):
+    """A helper for plotConvertedRZTReactor."""
+    for b in reactor.core.iterBlocks():
+        blockMidTh, blockMidR, blockMidZ = b.spatialLocator.getGlobalCoordinates(nativeCoords=True)
+        if (blockMidTh >= innerTheta) and (blockMidTh <= outerTheta):
+            if (blockMidR >= innerRadius) and (blockMidR <= outerRadius):
+                if (blockMidZ >= innerAxial) and (blockMidZ <= outerAxial):
+                    return b
+
+    raise ValueError(
+        "No block found between ({}, {}), ({}, {}), ({}, {})\nLast block had TRZ= {} {} {}".format(
+            innerTheta,
+            outerTheta,
+            innerRadius,
+            outerRadius,
+            innerAxial,
+            outerAxial,
+            blockMidTh,
+            blockMidR,
+            blockMidZ,
+        )
+    )
+
+
+def _getBlockColor(colConverter, colGenerator, blockColors, blockType):
+    """A helper for plotConvertedRZTReactor."""
+    nextColor = None
+    if blockType not in blockColors:
+        if "fuel" in blockType:
+            nextColor = "tomato"
+        elif "structure" in blockType:
+            nextColor = "lightgrey"
+        elif "radial shield" in blockType:
+            nextColor = "lightgrey"
+        elif "duct" in blockType:
+            nextColor = "grey"
+        else:
+            while True:
+                try:
+                    nextColor = next(colGenerator)
+                    colConverter.to_rgba(nextColor)
+                    break
+                except ValueError:
+                    continue
+
+    return nextColor
